requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.1.0
pydantic==2.5.0
pytest==9.0.1
pytest-watch==4.2.0
python-dotenv==1.0.0
